+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Reduced duplicate code - Connectivity, Motion, External and DIO share the same class
 - Removed NONE option from drop-down, NONE was workaround for a validation issue in Integration's Options and fixed as part of HA v0.112.0
 - Reduced code of camera's configuration in Integration's Options 
+- Improved generate configuration file process
 
 **Fixed bugs:**
 - Fixed - Generate configuration files